requirements.txt
django 
djangorestframework
drf_yasg
djangorestframework-simplejwt
markdown
django-filter
django-allauth
django-cors-headers
dj_rest_auth
psycopg2
pgvector
requests
kafka-python
python-dotenv
beautifulsoup4
sqlalchemy
openai
feedparser